APP GEMINI/requirements.txt

tensorflow==2.18.0
tf-keras==2.18.0
deepface==0.0.93
opencv-python
pillow
customtkinter

